+++ docs/changelog.md
@@ -27,6 +27,8 @@ Compiler changes:
  * Increased maximum common-audio-data size
  * Added `tad-compiler 64tass-enums` command
  * Added `tad-compiler 64tass-export` command
+ * Added hexadecimal segment suffix to `tad-compiler ca65-export`
+   (using the `-x` or `-X` command line options)
 
 65816 API changes:
  * Added a 64tass API (identical to the ca65 API)